+++ src/main.js
@@ -1,10 +1,10 @@
-import PointsModel from './model/point-model';
+import PointsModel from './model/points-model';
 import TripPresenter from './presenter/trip-presenter';
 import DestinationModel from './model/destination-model';
-import OffersModel from './model/offer-model';
+import OffersModel from './model/offers-model';
 import FilterModel from './model/filter-model';
 import FilterPresenter from './presenter/filter-presenter';
-import CreatePointButtonPresenter from './presenter/create-point-button-presenter';
+import NewPointButtonPresenter from './presenter/new-point-button-presenter';
 import { RenderPosition, render } from './framework/render';
 import TripInfoView from './view/trip-info-view';
 
@@ -24,7 +24,7 @@ const pointsModel = new PointsModel();
 const filterModel = new FilterModel();
 //console.log(filterModel);
 
-const createPointButtonPresenter = new CreatePointButtonPresenter({
+const createPointButtonPresenter = new NewPointButtonPresenter({
   container: container.tripInfo,
 });
 // console.log(createPointButtonPresenter);
@@ -38,7 +38,7 @@ const tripPresenter = new TripPresenter({
   createPointButtonPresenter,
 });
 
-render(new TripInfoView(pointsModel.getPoints(), destinationModel), container.tripInfo, RenderPosition.AFTERBEGIN);
+render(new TripInfoView(pointsModel.get(), destinationModel), container.tripInfo, RenderPosition.AFTERBEGIN);
 
 const filterPresenter = new FilterPresenter({
   filterContainer: container.filter,
