+++ src/model/offers-model.js
