+++ src/templates/new-point-button-template.js
